Move database to app/ dir
