+++ test/test_product_info.py
@@ -1,6 +1,6 @@
 import pytest
 from app.products.product_info import ProductInfoHandler
-from app.products.database import MOCK_PRODUCT_DATA
+from app.database import MOCK_PRODUCT_DATA
 @pytest.mark.prod_info
 
 class TestProductInfo:
